+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### FIX
 
-- Pay attention to correct return values.
+- [0007872](https://bugs.oxid-esales.com/view.php?id=7872): Pay attention to correct return values.
+- [0007874](https://bugs.oxid-esales.com/view.php?id=7874): Fix PayPal stateand country initialisation in purchase units factory
 
 ## [3.5.2] - 2025-12-05